Generalising the custom chat mode
